working to move all auth content into auth app
