+++ auth_app/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
+from .models import MyCustomUser
 
-# Register your models here.
+admin.site.register(MyCustomUser, UserAdmin)
