+++ backend/admin.py
@@ -1,6 +1,4 @@
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-from .models import MyCustomUser, FileUpload
+from .models import FileUpload
 
-admin.site.register(MyCustomUser, UserAdmin)
 admin.site.register(FileUpload)
