+++ backend/models.py
@@ -1,13 +1,13 @@
 from django.db import models
 from django.utils.timezone import now
-from django.contrib.auth.models import AbstractUser
+# from django.contrib.auth.models import AbstractUser
 
-class MyCustomUser(AbstractUser):
-    email = models.EmailField()
-    password = models.CharField(max_length=20)
+# class MyCustomUser(AbstractUser):
+#     email = models.EmailField()
+#     password = models.CharField(max_length=20)
 
-    def __str__(self):
-        return self.username
+#     def __str__(self):
+#         return self.username
 
 class FileUpload(models.Model):
     upload = models.FileField()
