+++ capstone/settings.py
@@ -128,4 +128,4 @@ USE_TZ = True
 
 STATIC_URL = '/static/'
 
-AUTH_USER_MODEL = 'backend.MyCustomUser'
+AUTH_USER_MODEL = 'auth_app.MyCustomUser'
