+++ capstone/urls.py
@@ -16,7 +16,8 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 from django.conf import settings
-from auth_app.views import LoginFormView, LogoutView, UploadView, SignupFormView
+from auth_app.views import LoginFormView, LogoutView, SignupFormView
+from backend.views import UploadView
 from django.conf.urls.static import static
 
 urlpatterns = [
